app/requirements.txt
Flask==2.3.3
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
SQLAlchemy==2.0.20
python-dotenv==1.0.0
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
google-cloud-pubsub==2.19.2 